Changed int to long long in validator (priority-research)

+++ priority-research/validator.cpp
@@ -7,7 +7,7 @@ const ll max_n = 1'000'000'000'000'000'000ll;
 int main(int argc, char** argv) {
     registerValidation(argc, argv);
     
-    ll N = inf.readInt(1, max_n, "N");
+    ll N = inf.readLong(1, max_n, "N");
     inf.readEoln();
 
     return 0;
